requirements.txt
openai
pypdf
tiktoken
gradio
python-dotenv